feat: challenge 4 done

+++ src/main/java/org/binar/repository/ShowtimeRepository.java
@@ -3,6 +3,6 @@ package org.binar.repository;
 import org.binar.model.Showtime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowtimeRepository extends JpaRepository<Showtime, Integer> {
+public interface ShowtimeRepository extends JpaRepository<Showtime, Long> {
 
 }
